renamed alot of doc imports or calls
